working on performance

+++ src/tools.h
@@ -26,18 +26,18 @@ typedef struct node
     struct node *next;
 } node_t, *list_t;
 
-int gen_rand(unsigned int, unsigned int);
+int gen_rand(unsigned short int, unsigned short int);
 
 SDL_Texture *create_txt(TTF_Font *, char *, SDL_Color);
 SDL_Texture *load_img(char *);
-SDL_Rect rect_from_texture(SDL_Texture *, unsigned int, unsigned int);
+SDL_Rect rect_from_texture(SDL_Texture *, unsigned short int, unsigned short int);
 
 list_t cons_empty(void);
 list_t cons(void *, list_t);
 void *head(list_t);
 list_t son(list_t);
 void print_list(list_t);
-void *pop_nth(list_t, unsigned int);
+void *pop_nth(list_t, unsigned short int);
 bool is_empty(list_t);
 void free_list(list_t);
